+++ simple-stack/src/test/java/com/zhuinden/simplestack/FlowTest.java
@@ -2,6 +2,7 @@ package com.zhuinden.simplestack;
 
 /*
  * Copyright 2013 Square Inc.
+ * Copyright 2017 Gabor Varadi
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.